Kafka+Flink/producer_requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
Authlib==1.5.2
avro==1.12.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
confluent-kafka==2.10.0
cryptography==44.0.2
decorator==5.2.1
fastavro==1.10.0
frozenlist==1.6.0
fsspec==2025.3.2
gcsfs==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
kafka-python==2.1.5
multidict==6.4.3
numpy==2.2.4
oauthlib==3.2.2
pandas==2.2.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0

